pyproject.toml
[project]
name = "fraud-detection-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.3",
    "boto3>=1.42.44",
    "matplotlib>=3.10.8",
    "pandas>=3.0.0",
    "psycopg2-binary>=2.9.11",
    "pyspark>=4.1.1",
    "python-dotenv>=1.2.1",
    "scikit-learn>=1.8.0",
    "sqlalchemy>=2.0.46",
    "torch>=2.10.0",
]
